change in directory structure
put all the jsp pages into web-inf and added pages to admin module also make multiple servlets to map all the pages of admin module with authentication
