+++ src/main/webapp/WEB-INF/web.xml
@@ -79,4 +79,84 @@
     <servlet-name>acceptUser</servlet-name>
     <url-pattern>/acceptUser</url-pattern>
   </servlet-mapping>
+  <servlet>
+    <description></description>
+    <display-name>UploadNotesServlet</display-name>
+    <servlet-name>UploadNotesServlet</servlet-name>
+    <servlet-class>in.NoteX.servlet.UploadNotesServlet</servlet-class>
+  </servlet>
+  <servlet-mapping>
+    <servlet-name>UploadNotesServlet</servlet-name>
+    <url-pattern>/UploadNotesServlet</url-pattern>
+  </servlet-mapping>
+  <servlet>
+    <description></description>
+    <display-name>mapUploadNotes</display-name>
+    <servlet-name>mapUploadNotes</servlet-name>
+    <servlet-class>in.NoteX.mapjsp.mapUploadNotes</servlet-class>
+  </servlet>
+  <servlet-mapping>
+    <servlet-name>mapUploadNotes</servlet-name>
+    <url-pattern>/mapUploadNotes</url-pattern>
+  </servlet-mapping>
+  <servlet>
+    <description></description>
+    <display-name>mapApproveUser</display-name>
+    <servlet-name>mapApproveUser</servlet-name>
+    <servlet-class>in.NoteX.mapjsp.mapApproveUser</servlet-class>
+  </servlet>
+  <servlet-mapping>
+    <servlet-name>mapApproveUser</servlet-name>
+    <url-pattern>/mapApproveUser</url-pattern>
+  </servlet-mapping>
+  <servlet>
+    <description></description>
+    <display-name>mapApproveNotesPyq</display-name>
+    <servlet-name>mapApproveNotesPyq</servlet-name>
+    <servlet-class>in.NoteX.mapjsp.mapApproveNotesPyq</servlet-class>
+  </servlet>
+  <servlet-mapping>
+    <servlet-name>mapApproveNotesPyq</servlet-name>
+    <url-pattern>/mapApproveNotesPyq</url-pattern>
+  </servlet-mapping>
+  <servlet>
+    <description></description>
+    <display-name>mapUploadPyq</display-name>
+    <servlet-name>mapUploadPyq</servlet-name>
+    <servlet-class>in.NoteX.mapjsp.mapUploadPyq</servlet-class>
+  </servlet>
+  <servlet-mapping>
+    <servlet-name>mapUploadPyq</servlet-name>
+    <url-pattern>/mapUploadPyq</url-pattern>
+  </servlet-mapping>
+  <servlet>
+    <description></description>
+    <display-name>mapUploadImp</display-name>
+    <servlet-name>mapUploadImp</servlet-name>
+    <servlet-class>in.NoteX.mapjsp.mapUploadImp</servlet-class>
+  </servlet>
+  <servlet-mapping>
+    <servlet-name>mapUploadImp</servlet-name>
+    <url-pattern>/mapUploadImp</url-pattern>
+  </servlet-mapping>
+  <servlet>
+    <description></description>
+    <display-name>mapUploadPlacementMaterial</display-name>
+    <servlet-name>mapUploadPlacementMaterial</servlet-name>
+    <servlet-class>in.NoteX.mapjsp.mapUploadPlacementMaterial</servlet-class>
+  </servlet>
+  <servlet-mapping>
+    <servlet-name>mapUploadPlacementMaterial</servlet-name>
+    <url-pattern>/mapUploadPlacementMaterial</url-pattern>
+  </servlet-mapping>
+  <servlet>
+    <description></description>
+    <display-name>mapApprovalDeleteUser</display-name>
+    <servlet-name>mapApprovalDeleteUser</servlet-name>
+    <servlet-class>in.NoteX.mapjsp.mapApprovalDeleteUser</servlet-class>
+  </servlet>
+  <servlet-mapping>
+    <servlet-name>mapApprovalDeleteUser</servlet-name>
+    <url-pattern>/mapApprovalDeleteUser</url-pattern>
+  </servlet-mapping>
 </web-app>
