+++ src/main/webapp/js/script.js
@@ -17,8 +17,22 @@ document.getElementsByClassName("card-img-top");
 
 
 
-let span =document.getElementsByClassName("col-6")[0];
-//span.style.color="red";
-span.textContent="vivek patel";
 
+
+function clearAllCookies() {
+    var cookies = document.cookie.split(";");
+
+    for (var i = 0; i < cookies.length; i++) {
+        var cookie = cookies[i];
+        var eqPos = cookie.indexOf("=");
+        var name = eqPos > -1 ? cookie.substr(0, eqPos) : cookie;
+        document.cookie = name + "=;expires=Thu, 01 Jan 1970 00:00:00 GMT";
+    }
+}
+
+
+window.onload = function() {
+    clearAllCookies();
+    /*alert("Cookies cleared!");*/
+};
